refactor(mcp-server): session-resolver moves out of mcp/ and becomes current-workspace
It stopped being an MCP concern the moment #1082 had `http-server.ts` call it
for the daemon's startup workspace. Left under `server/mcp/`, it made the
daemon's composition root import from the MCP adapter tree — which reads as
the daemon depending on a transport it does not use, and which is what the
ADR-0018 guard would say too once it can see this far (next commit).

`server/current-workspace.ts` is the wiring layer `app.ts` and
`http-server.ts` already live in, and both callers reach it from there.

The name goes with it. It resolves a WORKSPACE id and always did — nothing
about it is a session, and `.claude/rules/vocabulary.md` names Workspace as
the word for the thing a document lives in. Pure move plus rename: the
exported `ensureWorkspaceId` / `clearWorkspaceIdCache` and every line of
behavior are unchanged, and the eight importers follow.
