+++ packages/mcp-server/src/server/http-server.ts
@@ -12,9 +12,9 @@ import { PACKAGE_VERSION } from '../shared/package-version.js'
 import { WHITEBOARD_WS_PROTOCOL } from '../shared/ws-protocol.js'
 import { createApp } from './app.js'
 import { DIST_WEB_APP_DIR, getDataDir } from './config.js'
+import { ensureWorkspaceId } from './current-workspace.js'
 import { normalizeBindHost } from './daemon-auth-binding.js'
 import { getLogger } from './log.js'
-import { ensureWorkspaceId } from './mcp/session-resolver.js'
 import { getConnectionStats, handleWsUpgrade, setRuntimeTouchFn } from './routes/ws.js'
 import { authorizeWsUpgrade } from './routes/ws-auth.js'
 import { parseWsTargetFromRequestUrl } from './routes/ws-validation.js'
